blog/python_sdk_demo/requirements.txt
determined>=0.26.4
medmnist
PyYAML
